update • renamed function version and some adjustments

+++ DEMO_gray_func.m
@@ -1,13 +1,14 @@
-function DEMO_gray_lam2(lam_2_arg)
+function DEMO_gray_lam2(factor_arg, iter_outer,...
+    lam_2_arg, iter_inner, gau_var)
 
 %% Initialize parameters
-UPSAMP.factor = 4;  % magnification factor
-UPSAMP.iter = 3;  % times of iteration
+UPSAMP.factor = factor_arg;  % magnification factor
+UPSAMP.iter = iter_outer;  % upsampling feed-back iterations
 DECONV.lambda_1 = 0.01;  % lambda_1
 DECONV.lambda_2 = lam_2_arg;  % lambda_2
-DECONV.iter_max =10;  % max deconv iterations
+DECONV.iter_max = iter_inner;  % max deconv iterations
 GAU.size = 11;  % size of gaussian kernel
-GAU.var = 1;  % variance of gaussian kernel
+GAU.var = gau_var;  % variance of gaussian kernel
 IMG.name = 'words';
 IMG.read = ['images/', IMG.name, '_small.jpg'];
 
@@ -16,6 +17,15 @@ IMG.read = ['images/', IMG.name, '_small.jpg'];
 L = im2double(imread(IMG.read));
 L = rgb2gray(L);
 L = imbinarize(L);
+
+%% Show image L
+% Image L duplicatedly resized to show.
+L_show = pixeldup(L, UPSAMP.factor);
+close all;
+FIG_HANDLE.L = figure('Name', 'IMG - Low Resolution');
+movegui(FIG_HANDLE.L, 'west');
+imshow(L_show);
+title('Original image has been resized to display.');
 
 %% Upsample initially
 H_tilde = im2double(imresize(L, UPSAMP.factor, 'bicubic'));
@@ -54,17 +64,29 @@ figure(FIG_HANDLE.H);
 imshow(H_star);
 title(['After ', num2str(i), ' iterations']);
 
-mkdir(['results/', IMG.name]);
-IMG.write = ['results/', IMG.name, '/',...
-    [IMG.name, '_', IMG.appro,],...
+% Write in the disk
+IMG.folder = ['results/', IMG.name, '/'];
+IMG.src = [IMG.folder, IMG.name, '_lr.jpg'];
+IMG.write = [IMG.folder, IMG.name,...
+    ['_', IMG.appro,],...
     ['_', num2str(UPSAMP.factor), 'x'],...
     ['_', num2str(UPSAMP.iter), '_iter'],...
     ['_', num2str(GAU.size),'-', num2str(GAU.var), '_gau'],...
     ['_', IMG.other],...
     '.jpg'];
-imwrite(H_star, IMG.write);
-disp([newline,' ============>¡¾Image "',...
-    IMG.write, '" saved!¡¿',newline]);
+if ~isfile(IMG.src)
+    imwrite(L_show, IMG.src);
+    disp([newline,' ============>¡¾Image "',...
+        IMG.src, '" saved!¡¿',newline]);
+end
+if ~isfolder(IMG.folder)
+    mkdir(IMG.folder)
+end
+if ~isfile(IMG.write)
+    imwrite(H_star, IMG.write);
+    disp([newline,' ============>¡¾Image "',...
+        IMG.write, '" saved!¡¿',newline]);
+end
 
 %% Function on call
 function [I, low, gap] = simpnormimg(G)
